- Data now is float 32 and not float64 - saves a lot of memory - Each epoch the batches are generated from the full trajectory(that is already created). This allows each epoch be more independent from others - Removed unnesseray temp files

+++ Dataset.py
@@ -118,7 +118,7 @@ class CGTrajectoryDataSet(Dataset):
         dataSet, nCgDim = pt.CreateCoarseGrainedTraj(nDim, lenTrajFull, mWx, vHiddenStates, timeRes, semiCG=semiCG)
         kldEstimator, T, _, _, _, _ = pt.CalcKLDPartialEntropyProdRate(dataSet, nCgDim)
 
-        dataSet = torch.from_numpy(dataSet[:, 0])
+        dataSet = torch.from_numpy(dataSet[:, 0]).float()
         dataSet = torch.utils.data.TensorDataset(dataSet)
 
         # Define sampler
